requirements.txt
Django==2.2.14
django-cors-headers==3.4.0
djangorestframework==3.11.0
Pillow==7.2.0
pkg-resources==0.0.0
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
